+++ src/app/core/auth/containers/Login.tsx
@@ -8,7 +8,7 @@ import { loginThunk } from '@app/store/auth/thunk/authThunk';
 import { useAppDispatch } from '@app/store/hook/useAppDispatch';
 import { useAppSelector } from '@app/store/hook/useAppSelector';
 import { Button, Input } from '@shared/components/partials';
-import { validationRulesAuth } from '@shared/utils/validationRules';
+import { validationRulesAuth } from '@shared/constants/validationRules';
 
 import hideIcon from '@assets/icons/hide.svg';
 import showIcon from '@assets/icons/show.svg';
@@ -36,7 +36,7 @@ const Login = () => {
 
   const onSubmit = async (data: ILoginForm) => {
     try {
-      await dispatch(
+      const result = await dispatch(
         loginThunk({ email: data.email, password: data.password })
       );
       toast.success('Login successfully');
